+++ src/services/recoveryVaults/mutation.ts
@@ -1,0 +1,5 @@
+export default `mutation($input: [AddazureRecoveryVaultInput!]!) {
+  addazureRecoveryVault(input: $input, upsert: true) {
+    numUids
+  }
+}`;