+++ src/services/replicationAppliances/index.ts
@@ -1,0 +1,16 @@
+import { Service } from '@cloudgraph/sdk'
+import BaseService from '../base'
+import format from './format'
+import getConnections from './connections'
+import mutation from './mutation'
+import getData from './data'
+
+export default class AzureReplicationAppliance extends BaseService implements Service {
+  format = format.bind(this)
+
+  getConnections = getConnections.bind(this)
+
+  getData = getData.bind(this)
+
+  mutation = mutation
+}