+++ src/services/replicationCenters/mutation.ts
@@ -1,0 +1,5 @@
+export default `mutation($input: [AddazureReplicationCenterInput!]!) {
+  addazureReplicationCenter(input: $input, upsert: true) {
+    numUids
+  }
+}`;