+++ src/services/replicationNetworks/mutation.ts
@@ -1,0 +1,5 @@
+export default `mutation($input: [AddazureReplicationNetworkInput!]!) {
+  addazureReplicationNetwork(input: $input, upsert: true) {
+    numUids
+  }
+}`;