+++ src/services/synapseBigDataPools/mutation.ts
@@ -1,0 +1,5 @@
+export default `mutation($input: [AddazureSynapseBigDataPoolInput!]!) {
+  addazureSynapseBigDataPool(input: $input, upsert: true) {
+    numUids
+  }
+}`;